api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "API Gateway for the OpenTelemetry tracing demo"
requires-python = ">=3.12"
dependencies = [
    "flask",
    "requests",
    "opentelemetry-distro",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-flask",
    "opentelemetry-instrumentation-requests",
]
